feat: projects (M6)
- Projects entity with schemas, service, JSON CRUD and HTMX UI (list
  with status filter, detail, edit, delete-cascade). Tags stored as
  JSON-as-text per spec.
- Status workflow planning -> printing -> done | archived | abandoned
  with quick-action chips on the detail page; 'done' auto-stamps
  completed_at and clears it on transition back.
- project_models / project_filaments / project_items link tables wired
  with hydrated read endpoints (joined model thumbnail, filament color,
  inventory on-hand) so the UI doesn't N+1. Add-model link is
  idempotent on the composite PK.
- Migration 0003: project_items.qty_required + qty_consumed promoted to
  REAL so fractional quantities match the M4-followup inventory schema.
- BOM rollup endpoint + table on the project detail page using the
  CLAUDE.md formula (still_needed = qty_required - qty_consumed,
  still_to_buy = max(0, still_needed - on_hand)). Amber-highlighted
  rows when still_to_buy > 0.
- Cross-project shopping list: aggregates demand across active
  projects (status in planning, printing), subtracts on-hand once,
  lists project IDs needing each item. Surfaced as a dashboard section
  alongside a new "active projects" tile that replaces the M6
  placeholder.
- Verified end-to-end through Docker: created a project linked to a
  printer, attached a BOM item with on_hand=4 / required=20, confirmed
  /api/projects/1/bom and /api/shopping-list both return still_to_buy=16.

108 tests pass.

+++ src/maketrack/models/project.py
@@ -63,6 +63,8 @@ class ProjectItem(Base, TimestampMixin):
     inventory_item_id: Mapped[int] = mapped_column(
         ForeignKey("inventory_items.id", ondelete="RESTRICT"),
     )
-    qty_required: Mapped[int]
-    qty_consumed: Mapped[int] = mapped_column(default=0)
+    # Float to match inventory_items.quantity — track "1.5m of XT60 wire" or
+    # "0.25 kg of resin." Migration 0003 promotes the columns.
+    qty_required: Mapped[float]
+    qty_consumed: Mapped[float] = mapped_column(default=0.0)
     notes: Mapped[str | None] = mapped_column(default=None)
